+++ apps/client/src/app/auth/_state/server/hooks/index.ts
@@ -1,2 +1,1 @@
 export * from "./usePostSignUp";
-export * from "./useSignUpSubmitHandler";
